Code refactoring: - combined the code into one file - updating all lookup tables with new data points from the file before running the processing step - run the code from the terminal using parameters
