+++ Resources/agaf_map.txt
@@ -1,3 +1,3 @@
-7	חיפה והצפון
-8	דן והחוף
-9	ירושלים ונגב
+'7', '                         חיפה והצפון'
+'8', '                         דן והחוף'
+'9', '                         ירושלים ונגב'
