+++ Resources/iec_reply_map.txt
@@ -1,4 +1,4 @@
-'1', 'חיובית'
-'3', 'שלילית'
-'4', 'חיובית מוגבלת'
-'5', 'חיובית חלקית'
+'1', '          חיובית'
+'3', '          שלילית'
+'4', '          חיובית מוגבלת'
+'5', '          חיובית חלקית'
