+++ Resources/lms_yishuv_name_to_code_map.csv
@@ -1530,3 +1530,5 @@
 תל אל סאפי,1000047,,,,
 נופי פרת,1000048
 מידע חסר,1000049
+יתיר,1000050
+אתר 309,1000051
