+++ Resources/manual_map_of_yishuv_name.csv
@@ -91,3 +91,9 @@
 מת הכח נבי יושע,"מרכז כ""ח"
 מבואות גלבוע,מבואות הגלבוע
 זלפה,מעלה עירון
+מודיעין מכבים רעו,מודיעין-מכבים-רעות*
+נימינה-גבעת עדה,בנימינה-גבעת עדה*
+שדות יעקב איחוד,אשדות יעקב (איחוד)
+שדות יעקב מאוחד,אשדות יעקב (מאוחד)
+ית יצחק-שער חפר,בית יצחק-שער חפר
+ת אריה - עופרים,בית אריה-עופרים
