+++ Resources/napa_map.txt
@@ -1,9 +1,9 @@
-701	חיפה
-702	עמקים
-703	גליל וגולן
-801	דן
-802	שרון
-803	שפלה
-901	ירושלים
-902	נגב
-903	איילון
+'701', '                        חיפה'
+'702', '                        עמקים'
+'703', '                        גליל וגולן'
+'801', '                        דן'
+'802', '                        שרון'
+'803', '                        שפלה'
+'901', '                        ירושלים'
+'902', '                        נגב'
+'903', '                        איילון'
