+++ Resources/reg_number_map.txt
@@ -1,38 +1,39 @@
-'130', '   בור פיוי בינוני הליך תחרותי מכרז 2'
-'133', '    בור פיוי בינוני הליך תחרותי מכרז 3'
-'134', '    בור פיוי בינוני הליך תחרותי מכרז 4'
-'140', '    חיבור פיוי הליך תחרותי גגות מכרז 1'
-'141', '    חיבור פיוי הליך תחרותי גגות מכרז 2'
-'142', '    חיבור פיוי הליך תחרותי גגות מכרז 3'
-'160', '    ה.תחרותי גגות ללא מכרז'
-'161', '  ה.תחרותי קרקע ללא מכרז'
-'170', '    בשילוב קיבולת אגירה מכרז 1 PV יבור'
-'171', '    בשילוב קיבולת אגירה מכרז 2 PV יבור'
-'180', 'הליך תחרותי דו שימוש מכרז 1'
-'250', '    אסדרה תעריפית ברירת מחדל'
-'251', '    צריכה עצמית ללא תעריף'
-'260', '    מחלק היסטורי אסדרות חדשות 2018'
-'280', '    אסדרת יצרן ביו-גז'
-'290', '    אסדרה גז טבעי'
-'300', '    ה תעריפית עד 15 קוא ללא מונה ייצור'
-'301', '    ללא מונה ייצור ו   KVA 15 תעריפית עד'
-'305', '    אסדרה תעריפית עד 15 עם מונה ייצור'
-'308', '    אסדרה תעריפית מעל 15 קוא'
-'310', '    מתקנים עד 15'
-'311', '    מתקנים מעל 15 ועד 100'
-'312', '    חיבור מתקן תעריפי עד 630'
-'313', 'מסלול ירוק )מהיר( עד 10 קוא'
-'350', '    החלפת הספק עד להספק פחת'
-'351', '    הספק מעבר להספק הפחת ללא החלפת ממי'
-'352', '    הספק מעבר להספק הפחת עם תוספת ממיר'
-'378', '   מונה נטו מעל 15 קוא'
-'905', '    התקנת טורבינת רוח'
-'906', '    או   PV הגדלה/הוספה חיבור לצורך'
-'918', '  ביתי - הסדרה מ- 05/07/11 PV'
-'935', 'יצרנים פרטיים-טורבינות רוח'
-'940', '    יצרן פרטי'
-'941', '   מונה נטו מ.היסטורי מ- 02/4/13'
-'942', '   מונה נטו מ- 02/04/2013'
-'946', '   מונה נטו אחיד מ- 02/04/13'
-'964', '   PV 15/12/15-   מ.היסטורי עסקי מ'
-'969', '   PV 15/12/15 -   עסקי הסדרה מ'
+'130', '                            בור פיוי בינוני הליך תחרותי מכרז 2'
+'133', '                             בור פיוי בינוני הליך תחרותי מכרז 3'
+'134', '                             בור פיוי בינוני הליך תחרותי מכרז 4'
+'140', '                             חיבור פיוי הליך תחרותי גגות מכרז 1'
+'141', '                             חיבור פיוי הליך תחרותי גגות מכרז 2'
+'142', '                             חיבור פיוי הליך תחרותי גגות מכרז 3'
+'160', '                             ה.תחרותי גגות ללא מכרז'
+'161', '                           ה.תחרותי קרקע ללא מכרז'
+'170', '                             בשילוב קיבולת אגירה מכרז 1 PV יבור'
+'171', '                             בשילוב קיבולת אגירה מכרז 2 PV יבור'
+'180', '                         הליך תחרותי דו שימוש מכרז 1'
+'190', '                        מתקני חלוץ אגרו וולטאיים'
+'250', '                             אסדרה תעריפית ברירת מחדל'
+'251', '                             צריכה עצמית ללא תעריף'
+'260', '                             מחלק היסטורי אסדרות חדשות 2018'
+'280', '                             אסדרת יצרן ביו-גז'
+'290', '                             אסדרה גז טבעי'
+'300', '                             ה תעריפית עד 15 קוא ללא מונה ייצור'
+'301', '                             ללא מונה ייצור ו   KVA 15 תעריפית עד'
+'305', '                             אסדרה תעריפית עד 15 עם מונה ייצור'
+'308', '                             אסדרה תעריפית מעל 15 קוא'
+'310', '                             מתקנים עד 15'
+'311', '                             מתקנים מעל 15 ועד 100'
+'312', '                             חיבור מתקן תעריפי עד 630'
+'313', '                         מסלול ירוק )מהיר( עד 10 קוא'
+'350', '                             החלפת הספק עד להספק פחת'
+'351', '                             הספק מעבר להספק הפחת ללא החלפת ממי'
+'352', '                             הספק מעבר להספק הפחת עם תוספת ממיר'
+'378', '                            מונה נטו מעל 15 קוא'
+'905', '                             התקנת טורבינת רוח'
+'906', '                             או   PV הגדלה/הוספה חיבור לצורך'
+'918', '                           ביתי - הסדרה מ- 05/07/11 PV'
+'935', '                         יצרנים פרטיים-טורבינות רוח'
+'940', '                             יצרן פרטי'
+'941', '                            מונה נטו מ.היסטורי מ- 02/4/13'
+'942', '                            מונה נטו מ- 02/04/2013'
+'946', '                            מונה נטו אחיד מ- 02/04/13'
+'964', '                            PV 15/12/15-   מ.היסטורי עסקי מ'
+'969', '                            PV 15/12/15 -   עסקי הסדרה מ'
