+++ Resources/voltage_map.txt
@@ -1,2 +1,2 @@
-1	מתח נמוך
-2	מתח גבוה
+'1', '                      מתח נמוך'
+'2', '                      מתח גבוה'
